:hammer: Migrate clientApi impl to commonMain
